Add a fog_credentials feature check to allow parser functions to warn vs fail
This change adds a new puppet feature called fog_credentials which uses
the ENV FOG_RC value and default cloud.fog location to locate the fog
credentials that will be used by gardener.  If the credentials are not
found then we don't enable the feature, and the parsers can check
if the feature is enabled before they execute.

Change-Id: Icefa69cad50856800cdd2a40a396b1c18756a6f8